# Conflicts:
#	src/tests/mod.rs
#	src/window/adapters/direct_framebuffer.rs
#	src/window/adapters/double_buffered.rs
#	src/window/graphics.rs
#	src/window/windows/button.rs
#	src/window/windows/frame.rs
#	src/window/windows/label.rs
#	src/window/windows/list.rs
#	src/window/windows/menu.rs
#	src/window/windows/menu_bar_popup.rs
#	src/window/windows/multi_column_list.rs
#	src/window/windows/text_input.rs

+++ src/window/windows/menu_bar.rs
@@ -355,42 +355,42 @@ impl Window for MenuBar {
 
         // Draw menu bar background
         device.fill_rect(
-            bounds.x as usize,
-            bounds.y as usize,
-            bounds.width as usize,
-            MENU_BAR_HEIGHT as usize,
+            bounds.x,
+            bounds.y,
+            bounds.width,
+            MENU_BAR_HEIGHT as u32,
             self.bg_color,
         );
 
         // Draw bottom border
         device.fill_rect(
-            bounds.x as usize,
-            (bounds.y + MENU_BAR_HEIGHT as i32 - 1) as usize,
-            bounds.width as usize,
+            bounds.x,
+            bounds.y + MENU_BAR_HEIGHT as i32 - 1,
+            bounds.width,
             1,
             Color::new(180, 180, 180),
         );
 
         // Draw menu titles
         for (i, menu) in self.menus.iter().enumerate() {
-            let x = bounds.x as usize + menu.x;
-            let text_y = bounds.y as usize + (MENU_BAR_HEIGHT as usize - char_height) / 2;
+            let x = bounds.x + menu.x as i32;
+            let text_y = bounds.y + (MENU_BAR_HEIGHT as i32 - char_height as i32) / 2;
 
             // Highlight if hovered or open
             let is_active = self.hover_index == Some(i) || self.open_menu_index == Some(i);
             if is_active {
                 device.fill_rect(
                     x,
-                    bounds.y as usize,
-                    menu.width,
-                    MENU_BAR_HEIGHT as usize,
+                    bounds.y,
+                    menu.width as u32,
+                    MENU_BAR_HEIGHT as u32,
                     self.hover_bg_color,
                 );
             }
 
             // Draw title text
             device.draw_text(
-                x + MENU_TITLE_PADDING,
+                x + MENU_TITLE_PADDING as i32,
                 text_y,
                 &menu.title,
                 font.as_font(),
